Refactorings to pull out the Git DSL from the GitHub filr
