+++ source/platforms/platform.ts
@@ -34,7 +34,7 @@ export interface Platform {
   /** Pulls in the platform specific metadata for inspection */
   getPlatformDSLRepresentation: () => Promise<any>
   /** Pulls in the Code Review Diff, and offers a succinct user-API for it */
-  getReviewDiff: () => Promise<GitDSL>
+  getPlatformGitRepresentation: () => Promise<GitDSL>
   /** Creates a comment on the PR */
   createComment: (body: string) => Promise<any>
   /** Delete the main Danger comment */
