+++ source/runner/_tests/Executor.test.ts
@@ -8,11 +8,11 @@ describe("setup", () => {
     const platform = new FakePlatform()
     const exec = new Executor(new FakeCI({}), platform)
 
-    platform.getReviewDiff = jest.fn()
+    platform.getPlatformGitRepresentation = jest.fn()
     platform.getPlatformDSLRepresentation = jest.fn()
 
     await exec.dslForDanger()
-    expect(platform.getReviewDiff).toBeCalled()
+    expect(platform.getPlatformGitRepresentation).toBeCalled()
     expect(platform.getPlatformDSLRepresentation).toBeCalled()
   })
